+++ ethereum/eip712/eip712.go
@@ -8,6 +8,7 @@ import (
 	"reflect"
 	"strings"
 
+	sdkmath "cosmossdk.io/math"
 	"golang.org/x/text/cases"
 	"golang.org/x/text/language"
 
@@ -379,7 +380,7 @@ var (
 	hashType      = reflect.TypeOf(common.Hash{})
 	addressType   = reflect.TypeOf(common.Address{})
 	bigIntType    = reflect.TypeOf(big.Int{})
-	cosmIntType   = reflect.TypeOf(sdk.Int{})
+	cosmIntType   = reflect.TypeOf(sdkmath.Int{})
 	cosmosAnyType = reflect.TypeOf(&codectypes.Any{})
 )