+++ x/evm/migrations/v3/migrate_test.go
@@ -2,8 +2,9 @@ package v3_test
 
 import (
 	"fmt"
-	v3 "github.com/evmos/ethermint/x/evm/migrations/v3"
 	"testing"
+
+	v3 "github.com/evmos/ethermint/x/evm/migrations/v3"
 
 	"github.com/stretchr/testify/require"
 
@@ -24,7 +25,7 @@ func TestMigrateStore(t *testing.T) {
 	tEvmKey := sdk.NewTransientStoreKey(fmt.Sprintf("%s_test", types.StoreKey))
 	ctx := testutil.DefaultContext(evmKey, tEvmKey)
 	paramstore := paramtypes.NewSubspace(
-		encCfg.Marshaler, encCfg.Amino, evmKey, tEvmKey, "evm",
+		encCfg.Codec, encCfg.Amino, evmKey, tEvmKey, "evm",
 	).WithKeyTable(v3types.ParamKeyTable())
 
 	params := v3types.DefaultParams()
@@ -39,7 +40,7 @@ func TestMigrateStore(t *testing.T) {
 	require.NotNil(t, preMigrationConfig.MergeForkBlock)
 
 	paramstore = paramtypes.NewSubspace(
-		encCfg.Marshaler, encCfg.Amino, evmKey, tEvmKey, "evm",
+		encCfg.Codec, encCfg.Amino, evmKey, tEvmKey, "evm",
 	).WithKeyTable(types.ParamKeyTable())
 	err := v3.MigrateStore(ctx, &paramstore)
 	require.NoError(t, err)