+++ types/int.go
@@ -5,7 +5,8 @@ import (
 	math "math"
 	"math/big"
 
-	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkmath "cosmossdk.io/math"
+
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )
 
@@ -21,11 +22,11 @@ func SafeInt64(value uint64) (int64, error) {
 }
 
 // SafeNewIntFromBigInt constructs Int from big.Int, return error if more than 256bits
-func SafeNewIntFromBigInt(i *big.Int) (sdk.Int, error) {
+func SafeNewIntFromBigInt(i *big.Int) (sdkmath.Int, error) {
 	if !IsValidInt256(i) {
-		return sdk.NewInt(0), fmt.Errorf("big int out of bound: %s", i)
+		return sdkmath.NewInt(0), fmt.Errorf("big int out of bound: %s", i)
 	}
-	return sdk.NewIntFromBigInt(i), nil
+	return sdkmath.NewIntFromBigInt(i), nil
 }
 
 // IsValidInt256 check the bound of 256 bit number